+++ README.adoc
@@ -4,22 +4,18 @@
 = SpringXMLConfig
 
 == Co to jest?
-W tym repozytorium znajdziesz przykładowe zadania oraz prezentacje przedstawiającą konfigurację aplikacji springowej przez pliki XML. +
+W tym repozytorium znajdziesz przykładowe zadania oraz prezentację przedstawiającą konfigurację aplikacji springowej przez pliki XML. +
 Zagadnienia:
 
  * Podstawy konfiguracji xmlowej - setter, konstruktor, fabryka ziaren
-
  * Łączenie kilku konfiguracji
-
  * Zasięg ziaren
-
  * Tworzenie prostej aplikacji z wykorzystaniem kontrolera, serwisu i repozytorium
-
  * Dobre praktyki
 
 
 === Instrukcja:
-Aby wykonać wszystkie ćwiczenia zacznij kolejno od pliku
+Aby wykonać wszystkie ćwiczenia zacznij od pliku
 
 
 [source, java]
@@ -29,7 +25,7 @@ package-info.java
 
 w pakiecie [yellow]#_a_wstrzykiwanie_zależności_# i podążaj za opisaną tam instrukcją.
 
-Kiedy skończysz opisane tam zadania, przejdź do pakietu [yellow]#_b_scope_i_import_# i również zacznij podążąć za instrukcją w:
+Kiedy skończysz opisane tam zadania, przejdź do pakietu [yellow]#_b_scope_i_import_# i również zacznij podążać za instrukcją znajdującą się w pliku:
 [source, java]
 ----
 package-info.java
