+++ src/main/java/ja/workshop/spring/części/konfiguracja/xml/b_scope_i_import/Main.java
@@ -1,4 +1,4 @@
-package wrup.spring.b_scope_i_import;
+package ja.workshop.spring.części.konfiguracja.xml.b_scope_i_import;
 
 
 import org.springframework.context.ApplicationContext;
