+++ src/main/java/ja/workshop/spring/części/konfiguracja/xml/c_mvc/KomisKontroler.java
@@ -1,4 +1,4 @@
-package wrup.spring.c_mvc;
+package ja.workshop.spring.części.konfiguracja.xml.c_mvc;
 
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.AbstractController;
