+++ src/main/java/ja/workshop/spring/części/konfiguracja/xml/c_mvc/silnik/Benzyna.java
@@ -1,7 +1,7 @@
-package wrup.spring.c_mvc.silnik;
+package ja.workshop.spring.części.konfiguracja.xml.c_mvc.silnik;
 
 /**
- * Ziarno springowe wstrzykiwane do klasy Samochód {@link wrup.spring.c_mvc.Samochód Samochód}
+ * Ziarno springowe wstrzykiwane do klasy {@link ja.workshop.spring.części.konfiguracja.xml.c_mvc.Samochód Samochód}.
  *
  * @author Wiktor Rup
  */
