+++ src/main/java/ja/workshop/spring/części/konfiguracja/xml/c_mvc/silnik/Silnik.java
@@ -1,4 +1,4 @@
-package wrup.spring.c_mvc.silnik;
+package ja.workshop.spring.części.konfiguracja.xml.c_mvc.silnik;
 
 /**
  * @author Wiktor Rup
