--- src/main/java/wrup/spring/a_wstrzykiwanie_zależności/silnik/Benzyna.java
@@ -1,12 +1,0 @@
-package wrup.spring.a_wstrzykiwanie_zależności.silnik;
-
-/**
- * Ziarno springowe wstrzykiwane do klasy Samochód {@link wrup.spring.a_wstrzykiwanie_zależności.Samochód Samochód}
- *
- * @author Wiktor Rup
- */
-public class Benzyna implements Silnik {
-    public String start() {
-        return "Pyr pyr pyr...";
-    }
-}
