--- src/main/java/wrup/spring/b_scope_i_import/silnik/OlejNapędowy.java
@@ -1,13 +1,0 @@
-package wrup.spring.b_scope_i_import.silnik;
-
-/**
- * Ziarno spingowe wstrzykiwane do klasy Samochód {@link wrup.spring.b_scope_i_import.Samochód Samochód}
- *
- * @author Wiktor Rup
- */
-public class OlejNapędowy implements Silnik {
-
-    public String start() {
-        return "Wrumm!";
-    }
-}
